+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   extends: './node_modules/@surmon-china/abc-factory/preset/eslintrc/vue.typescript',
   rules: {
-    'vue/max-attributes-per-line': [0]
+    'vue/max-attributes-per-line': 0,
+    'vue/order-in-components': 0
   }
 }